Fix ambiguity in conjuctions and tuple literals.
This change removes the possibility of using tuple literals in the first part
of the conjuction conditional. This removes the ambiguity in the syntax.

For example, consider this valid conditional:

"Applejack has more than one and Pinkie Pie has more than one"

Without this change, "one and Pinkie Pie" will be parsed as a tuple literal,
which then causes the parser to fail on the "has" keyword.